refactor: clean up snake movement logic
- split snake movement into separate files
- restructure project files
- update changelog file
